refactor: change GetRenewalInfo signature
